+++ src/tests/UserInputTests.java
@@ -9,8 +9,8 @@ class UserInputTests {
 	@Test
 	void coordinatesAreValid() {
 		UserInput testingCoordinates = new UserInput();
-		int[] coordinates1 = {-90, -180};
-		int[] coordinates2 = {90, 180};
+		double[] coordinates1 = {-90, -180};
+		double[] coordinates2 = {90, 180};
 		boolean correct1 = testingCoordinates.validateCoordinates(coordinates1);
 		boolean correct2 = testingCoordinates.validateCoordinates(coordinates2);
 		assertTrue(correct1);
@@ -20,10 +20,10 @@ class UserInputTests {
 	@Test
 	void coorindatesAreInvalid() {
 		UserInput testingCoordinates = new UserInput();
-		int[] coordinates1 = {-91, 0};
-		int[] coordinates2 = {91, 0};
-		int[] coordinates3 = {0, 181};
-		int[] coordinates4 = {0, -181};
+		double[] coordinates1 = {-91, 0};
+		double[] coordinates2 = {91, 0};
+		double[] coordinates3 = {0, 181};
+		double[] coordinates4 = {0, -181};
 		boolean incorrect1 = testingCoordinates.validateCoordinates(coordinates1);
 		boolean incorrect2 = testingCoordinates.validateCoordinates(coordinates2);
 		boolean incorrect3 = testingCoordinates.validateCoordinates(coordinates3);
